# artifacts/handoff-w4-round-1/provider-artifacts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}": {
      "get": {
        "operationId": "OrganizationsController_get",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/v1/organizations/{organizationId}/wallet/providers": {
      "get": {
        "operationId": "WalletController_providers",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/wallet-status": {
      "get": {
        "operationId": "WalletController_programStatus",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/wallet/reconcile": {
      "post": {
        "operationId": "WalletController_reconcile",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/wallet-sync/{jobId}": {
      "get": {
        "operationId": "WalletController_reconciliationStatus",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/v1/customer/wallet/apple/pass": {
      "get": {
        "operationId": "WalletController_applePass",
        "parameters": [
          {
            "name": "tenant",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/v1/customer/wallet/google/add-action": {
      "post": {
        "operationId": "WalletController_googleAction",
        "parameters": [
          {
            "name": "tenant",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
    "/v1/organizations/{organizationId}/staff-devices": {
      "get": {
        "operationId": "MerchantStaffDeviceController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
    "/v1/organizations/{organizationId}/device-pairing-sessions": {
      "post": {
        "operationId": "MerchantStaffDeviceController_createPairing",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
    "/v1/organizations/{organizationId}/device-pairing-sessions/{sessionId}": {
      "get": {
        "operationId": "MerchantStaffDeviceController_getPairing",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
    "/v1/organizations/{organizationId}/device-pairing-sessions/{sessionId}/cancel": {
      "post": {
        "operationId": "MerchantStaffDeviceController_cancelPairing",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
    "/v1/organizations/{organizationId}/staff-devices/{deviceId}/revoke": {
      "post": {
        "operationId": "MerchantStaffDeviceController_revoke",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
    "/v1/organizations/{organizationId}/staff-devices/{deviceId}/mark-compromised": {
      "post": {
        "operationId": "MerchantStaffDeviceController_markCompromised",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "MerchantStaffDevice"
        ]
      }
    },
...
    "/v1/staff/device-context": {
      "get": {
        "operationId": "StaffDevicePairingController_context",
        "parameters": [
          {
            "name": "x-waflo-device-id",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StaffDevicePairing"
        ]
      }
    },
...
  },
  "info": {
    "title": "Waflo Platform API",
    "description": "Authoritative Waflo W4 API for merchant administration and signed Staff devices.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "waflo_session"
      }
    },
...
  }
}